Simplify 'child view' a.k.a. our current FormSection object
Instead of having a child record's fields values loaded into the 
FormSection object, views only get the form structure from FormSections
and pull the values from the corresponding Child object. 

Motivation for change:

FormSection used to be a 'view model' (was previously named ChildView)
which was created based on the Templates object, and pulled field
values to be rendered in the view from a Child object.  The idea was
that the views wouldn't depend on Child objects, only ChildView/FormSection
objects which contained everything they needed.  Whilst this idea is sound,
it has prevented us from achieving a clean method of loading FormSections
at runtime from the database (since the class required injection of child 
properties).