Hacky fix for redirect loop
We got an "UnhandledModalError" or something like that from
capybara_features/manage_users.feature:20. Adding a sleep and clicking
the "Yes" button manually revealed a dialog asking if we were ok with
our data being redirected. This is confusing because we expect the
browser to do a GET and NOT resend any data.

In manual testing after seeing that, we saw in the Chromium firebuggy
thing that we seemed to be in a redirect loop. Chromium reported that
the first request was a PUT and subsequent requests were GETs, but the
Rails logs showed all the requests as PUTs.

The modal dialog didn't appear in manual testing, only via Selenium.

This workaround is to only redirect after update for non-XHR requests.

That's a crappy workaround. We'd like to understand why the server is
seeing the redirected requests as PUTs even though the browser reports
that they're GETs. Here's a guy who saw the same thing in March 2012:
http://www.ruby-forum.com/topic/3907238