[#1814] Mark - Remove WIP'd tests and steps
All the tests in reorder_fields.feature are marked with @wip and do not
currently pass. The functionality has clearly changed since they were
written, since there are no "up" and "down" arrows any more. Given this,
it seems like the tests might as well be removed for now; they have been
around for a while and do not seem to have been fixed and are just
adding noise to the code base.