[#1814] Mark - Move step to more logical place
The step was for a different page to the steps file it had been placed
in. Also updated to use the form_section_edit_page. Ideally it would
use a better selector and take the text on the page as an argument,
but this is a simple way of moving the step without adding that
complexity (and the step is only used in one place anyway).