@ctumwebaze @mwangiann rapidftr/tracker#152 processing enquiries on the server
This fix changes the way enquiries were processed on the server. It introduces a criteria
generation step where a criteria is build from the fields on the form sections and the enquiry data
before the enquiry is actually saved.

previously the criteria was specified as part of the enquiry data when the server received it but that is not
necessary anymore since we generate it from the enquiries fields.