removes score_threshold system variable.
when the enquiries are turned off, the SCORE_THRESHOLD system variable
is also not necessary and the user should not see it when they go to the
system variable administration page.

this fix extends the rake task to remove the SCORE_THRESHOLD variable
when enquiries are disabled or add it when the enquiries are enabled.