create enable enquiries variable on disable
when the enable enquiries variable doesn't exist and one runs the
app:enquiries:disable task, it's not created. This leaves the app in an
inconsistent state with the enquiries for deleted but the value of the
variable cannot be determined there fore the system assumes enquiries
are still turned on.

this fix ensures the enquiries variable is created even when the disable
enquiries rake task is run.