turn off enquiry permissions #252
when the enquiries feature is disable, the permissions class should not
return permissions for the enquiries and potential matches entities and should not be available
for the administrator to add or remove from roles.

the permission model is extended to check whether the enquiries feature
is on for the public facing methods that return permissions available to
the system. If the enquiries feature is on, the permissions will be
returned otherwise they wont.