making system variables uneditable [skip ci]
some system variables should not be editable by the user by the user
should still be able to see the value of that variable.

to achieve that, an attribute user_editable was added to the system
variable model to store a value indicating whether the user can edit the
value of it or not and retrofitted the system variables view to account
for the value of that attribute.