sets user_editable flag for enquiries variable [skip ci]
when the enable or disable tasks for the enquiries are run, the
user_editable property of the enable_enquiries system variable is also
set. This ensures that when the task is run against apps that already
have the enable_enquiries variable created will also have the right
state of the variable as expected by the app.